updated nuspec+chocolatey, integrated winbison (v2.x)
Including both in a single branch was originally brought up in #16, also using the GitHub release download instead of the SF one.
Even if we do not include bison2 in master those files actually don't depend on the branch itself but on a released version and therefore can be kept identical and changes simply merged over.