Fix CI, include all licenses, simplify CMake (#102)
* AppVeyor: Rewrite to only use CMake commands

* Chocolatey: Update version to current release

* Readmes: Point to GitHub instead of Sourceforge

* Readme: CMake is now mandatory

Since 1176da2adaf972b2216b1ac8dd611e128fe80bc5

* Expand .gitignore

* nit: line-endings

This commit is basically just a git rebase leftover. Originally updated CMake versions in the sub-projects which is already merged now.

* Readmes: VS2017 is the minimum supported version

* CPack: Actually install all license files

This change, plus the changes applied in #84 fixes #94

* CMake: $PROJECT_NAME is built-in

No need to define this manually

https://cmake.org/cmake/help/v3.10/variable/PROJECT_NAME.html

* CMake: The entire project is C only

This is an implicit performance optimization as CMake does not have to check for C++ capabilities in the setup

* CPack: Rename default dev package from master to dev-PROJECT_VERSION

* CMake: Simplify setting default CMAKE_BUILD_TYPE

* CMake: More explicit syntax to specific project language

* Readme: Simplify

* AppVeyor: Include "vs" in artifact names